Removed duped methods, migrated access to global vars to storeChecker, renamed Checker to ConnectionChecker
