+++ src/Constants/States.js
@@ -1,15 +1,18 @@
 /**
+ * @access public
  * @description Contains connection states
  * @constant {Object}
  */
 const ConnectionState = {
   /**
+   * @access public
    * @description Connected network state
    * @constant {ConnectionState}
    */
   CONNECTED: 'CONNECTED',
 
   /**
+   * @access public
    * @description Disconnected network state
    * @constant {ConnectionState}
    */
