+++ src/main.js
@@ -1,14 +1,23 @@
 import ConnectionState from 'Constants/States'
 import ConnectionEvent from 'Constants/Events'
-import { getConnectionState, changeTimeout, changeInterval, changeInternetResource, isCheckerActive, startChecker, stopChecker, checkOnDemand } from 'Checker'
-
+import { startChecker, stopChecker, checkOnDemand } from 'ConnectionChecker'
+import { getCheckerInstance, getConnectionState, isCheckerActive, getFetchTimeout, getIntervalTime, getInternetResource, setCheckerInstance, setConnectionState, setFetchTimeout, setIntervalTime, setInternetResource } from 'storeChecker'
+// FROM CONSTANTS
 exports.ConnectionState = ConnectionState
 exports.ConnectionEvent = ConnectionEvent
-exports.getConnectionState = getConnectionState
-exports.changeTimeout = changeTimeout
-exports.changeInterval = changeInterval
-exports.changeInternetResource = changeInternetResource
-exports.isCheckerActive = isCheckerActive
+// FROM CHECKER
 exports.startChecker = startChecker
 exports.stopChecker = stopChecker
 exports.checkOnDemand = checkOnDemand
+// FROM STORE
+exports.getCheckerInstance = getCheckerInstance
+exports.getConnectionState = getConnectionState
+exports.isCheckerActive = isCheckerActive
+exports.getFetchTimeout = getFetchTimeout
+exports.getIntervalTime = getIntervalTime
+exports.getInternetResource = getInternetResource
+exports.setCheckerInstance = setCheckerInstance
+exports.setConnectionState = setConnectionState
+exports.setFetchTimeout = setFetchTimeout
+exports.setIntervalTime = setIntervalTime
+exports.setInternetResource = setInternetResource
